refactor(ui): split the rewritten dropdown
