+++ ui/src/pages/Logs.tsx
@@ -9,7 +9,7 @@ import { ThemeContext } from "@/contexts/ThemeContext";
 
 import Button from "@/components/core/Button";
 import Input from "@/components/core/Input";
-import Dropdown from "@/components/core/Dropdown";
+import Dropdown from "@/components/core/DropdownOld";
 import RepositoriesDropdown from "@/components/dropdowns/RepositoriesDropdown";
 import DateDropdown from "@/components/dropdowns/DateDropdown";
 import Toggle from "@/components/core/Toggle";
